% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{corG}
\alias{corG}
\title{Gamma correlation}
\usage{
corG(d)
}
\arguments{
\item{d}{two ordinal random variables distribution}
}
\value{
vealue of Gamma correlation
}
\description{
calculating Gamma correlation
}
\examples{
\dontrun{
d <- matrix(ceiling(runif(12,1,12)), 3, 4)
corG(d)
}
}
