% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{corK}
\alias{corK}
\title{Kendall correlation}
\usage{
corK(x, y)
}
\arguments{
\item{x}{the first ordinal data(vector)}

\item{y}{the second ordinal data(vector)}
}
\value{
value of Kendall correlation
}
\description{
Calculating Kendall correlation
}
\examples{
\dontrun{
x <- runif(10,1,10);y <- runif(10,1,10)
x <- ceiling(x); y <- ceiling(y)
corK(x, y)
}
}
