% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{corS}
\alias{corS}
\title{Spearman correlation}
\usage{
corS(x, y)
}
\arguments{
\item{x}{the first ordinal data(vector)}

\item{y}{the second ordinal data(vector)}
}
\value{
value of Spearman correlation
}
\description{
Calculating Spearman correlation
}
\examples{
\dontrun{
x <- runif(10,1,10);y <- runif(10,1,10)
x <- ceiling(x); y <- ceiling(y)
corS(x, y)
}
}
