\name{StatComp20031-package}
\alias{StatComp20031-package}
\alias{StatComp20031}
\docType{package}
\title{
  Final project for Statistical Computing
}
\description{
  This package includes several functions to calculate ordinal correlaation and all homework as the final project for Statistical Computing course.
}
\author{
20031

Maintainer: 20031 <3256829518@qq.com>
}

\keyword{ package }


